+++ Layer_Fadecandy.h
@@ -7,10 +7,10 @@
 
 class SMLayerFadecandy : public SM_Layer {
     public:
-        SMLayerFadecandy(fcBuffers * newfcbuffers, uint8_t width, uint8_t height);
+        SMLayerFadecandy(fcBuffers * newfcbuffers, uint16_t width, uint16_t height);
         void frameRefreshCallback();
-        void fillRefreshRow(uint8_t hardwareY, rgb48 refreshRow[]);
-        void fillRefreshRow(uint8_t hardwareY, rgb24 refreshRow[]);
+        void fillRefreshRow(uint16_t hardwareY, rgb48 refreshRow[]);
+        void fillRefreshRow(uint16_t hardwareY, rgb24 refreshRow[]);
 
     private:
         inline uint32_t calculateFcInterpCoefficient();
